src/app/bookings/bookings.component.spec.ts
import {ComponentFixture, getTestBed, TestBed} from '@angular/core/testing';

import { BookingsComponent } from './bookings.component';
import {HttpClientTestingModule, HttpTestingController} from '@angular/common/http/testing';
import {RouterTestingModule} from '@angular/router/testing';

describe('BookingsComponent', () => {
  let component: BookingsComponent;
  let fixture: ComponentFixture<BookingsComponent>;
  let httpMock: HttpTestingController;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ BookingsComponent ],
      imports: [
        RouterTestingModule,
        HttpClientTestingModule
      ]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(BookingsComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
    httpMock = getTestBed().get(HttpTestingController);
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should return list of auditoriums', () =>{
     let lists = component.showBookings()
    expect(lists).toBeUndefined();
  });


  it('should logout', () => {
    component.logout();
    expect(component).toBeDefined();
  })

});
